Warning, compile & script binding fixes related to MathDefs.h. Use multiple template parameters where necessary to allow e.g. Min() or Max() with mixed int and unsigned values. Silence repeated warnings in MathDefs.h on MSVC.

+++ Source/Urho3D/Graphics/DecalSet.cpp
@@ -258,7 +258,7 @@ void DecalSet::SetMaterial(Material* material)
 void DecalSet::SetMaxVertices(unsigned num)
 {
     // Never expand to 32 bit indices
-    num = (unsigned)Clamp((int)num, MIN_VERTICES, MAX_VERTICES);
+    num = (unsigned)Clamp(num, MIN_VERTICES, MAX_VERTICES);
 
     if (num != maxVertices_)
     {
